+++ ESFramework/Foundation/NSNumberFormatter+ESAdditions.h
@@ -1,0 +1,18 @@
+//
+//  NSNumberFormatter+ESAdditions.h
+//  ESFramework
+//
+//  Created by Elf Sundae on 2019/04/24.
+//  Copyright © 2019 www.0x123.com. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface NSNumberFormatter (ESAdditions)
+
+/**
+ * The shared NSNumberFormatter instance.
+ */
++ (instancetype)defaultFormatter;
+
+@end